fix: duplicate dialog
