+++ packages/trpc/server/routers/viewer/eventTypes/duplicate.schema.ts
@@ -1,7 +1,7 @@
 import type { z } from "zod";
 
-import { EventTypeDuplicateInput } from "./types";
+import { EventTypeDuplicateInput } from "@calcom/prisma/zod/custom/eventtype";
 
-export const ZDuplicateInputSchema = EventTypeDuplicateInput.strict();
+export const ZDuplicateInputSchema = EventTypeDuplicateInput;
 
 export type TDuplicateInputSchema = z.infer<typeof ZDuplicateInputSchema>;
